fix some errors

+++ database/migrations/2025_01_30_113014_create_sections_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('sections', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('year_level');
+            $table->unsignedTinyInteger('year_level');
             $table->foreignId('course_id')->constrained()->onDelete('cascade');
             $table->timestamps();
             $table->unique(['name', 'year_level', 'course_id']);
